+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Bug Fixes & Changes
 - Milkweed can now be sheared with a Dispenser
 - Glory of the Snow can now be sheared with a Dispenser
 - Prickly Pear's Damage Source has been added to the bypasses_armor tag
+- Tumbleweed entities will now drop themselves if killed by any entity with Silk Touch, instead of only doing so when killed by a player
 
 - Fixed Fireflies with the Nectar easter egg not properly rendering their glowing overlay
 - Fixed Firefly Bottles with the Nectar easter egg not using the Nectar texture